Extracted Buy Pro button into separate view
